Integrated [WSTX-234] from 4.0; restores WstxInputProperties.NORMALIZE_LFS, adds WstInputProperties.TREAT_CHAR_REFS_AS_ENTS

+++ wstx/trunk/src/test/stax2/typed/ReaderBinaryTestBase.java
@@ -739,6 +739,6 @@ public abstract class ReaderBinaryTestBase
     }
 
     final static String displayByte(byte b) {
-        return "0x"+Integer.toHexString((int) b & 0xFF);
+        return "0x"+Integer.toHexString(b & 0xFF);
     }
 }
